[BE] refactor: ìAnalsysResult compatibilityScore 타입 변경에따른 전체 수정 (#91)
* [BE] refactor: ìAnalsysResult compatibilityScore 타입 변경에따른 전체 수정

* fix: AnalysisService AI 응답 파싱 검증 강화

  - AI 응답 형식 오류 시 예외 처리 개선
  - 잘못된 입력값으로 인한 런타임 오류 방지

+++ backend/src/main/java/com/devmatch/backend/domain/analysis/entity/AnalysisResult.java
@@ -9,6 +9,7 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.OneToOne;
 import jakarta.persistence.Table;
+import java.math.BigDecimal;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -31,7 +32,7 @@ public class AnalysisResult {
   private Application application;
 
   @Column(name = "compatibility_score", precision = 5, scale = 2, nullable = false)
-  private double compatibilityScore;
+  private BigDecimal compatibilityScore;
 
   @Column(columnDefinition = "TEXT", nullable = false)
   private String compatibilityReason;
